### 2017-09-06 #149 @MahShaaban
Submission: cRegulome
Obtains a database of pre-calculated microRNA and  transcription facor-gene correlations in cancer. In addition, the package  defines methods for handling and visualizing the data.

```
Package: cRegulome
Title: Obtain and Visualize regulome-gene Expression Correlations in Cancer
Version: 0.99.0
Author: Mahmoud Ahmed
Maintainer: Mahmoud Ahmed <mahmoud.s.fahmy@students.kasralainy.edu.eg>
Description: Obtain a database of pre-calculated microRNA and
  transcription facor-gene correlations in cancer. In addition, the package
  define basic methods for handling and visualizing the data.
Imports: DBI,
   RCurl,
   RSQLite,
   R.utils,
   UpSetR,
   VennDiagram,
   dbplyr,
   ggjoy,
   reshape2,
   grid
Suggests: BiocStyle,
    covr
biocViews: GeneExpression,
    Transcription
```

https://github.com/MahShaaban/cRegulome

- Please indicate which category or categories from our [package fit policies]
Data retrieval, because the package obtains a local database of Cistrome Cancer and miRCancerdb databases

Biologists and computational biologists with little knowledge of are and interested in investigating gene expression regulation in cancer.
