### 2017-09-13 #149 @karthik
Thanks for the submission! Looking for reviewers now. I also have two comments about one vignette and one test below.

### Editor checks:

---
Here are some suggestions from `goodpractice`. I would suggest increasing test coverage as the most important issue.

── GP cRegulome ─────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 63% of code lines are covered by test cases.

    R/get_data.R:43:NA
    R/get_data.R:48:NA
    R/get_data.R:53:NA
    R/get_data.R:59:NA
    R/get_data.R:60:NA
    ... and 140 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific functions you need.
  ✖ fix this R CMD check ERROR: Packages suggested but not available: ‘BiocStyle’ ‘rmarkdown’ The suggested packages are required for a complete check. Checking can be attempted without them by setting the environment variable _R_CHECK_FORCE_SUGGESTS_ to a false value.
    See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’ manual.
────────────────────────────────────

Minor comments:
- I would recommend ignoring `.DS_Store` in your [global gitignore](https://help.github.com/articles/ignoring-files/) (or just for this project at the least).

- Consider replacing `Rcurl` with `curl`.
