### 2017-09-13 #149 @MahShaaban
Thanks @karthik. I am looking forward.
Regarding your comments:
+ I put the `.DS_Store` to the `.gitignore` file
+ I am using `Rcurl` only because of the function `url.exists`. I am not aware of an alternative in `curl`.
+ When I run either `devtools::check()` or `goodpractice::gp()` I don't get any errors. I am not sure what to do since I can't reproduce the same `goodpractice` report.
+ Regarding the package coverage, I am working on it. Are you looking for a 100% coverage?!
