### 2017-09-13 #149 @karthik
For a replacement to `url.exists`, there are also options in `httr`
e.g. `http_error("http://www.google.com")`

> Are you looking for a 100% coverage?!

No! As reasonable an amount of coverage as possible. We realize that 100% coverage is hard to achieve for most packages.

I'm currently seeking reviewers and have sent out requests. I will update the thread when I'm ready to assign reviewers.
