### 2017-09-20 #149 @PeteHaitch
Hi @MahShaaban,

I've just started to look at **cRegulome**. I first tried running `devtools::check()` but I get an error during the creation of the vignette (see below). I note that **cRegulome** is building fine on Travis. The error occurs on my machine in the call `upset(ob_tf, study = 'ACC')` in the vignette:

```R
upset(ob_tf, study = 'ACC')
#> Error in mutate_impl(.data, dots) :
#> Evaluation error: object 'x' not found.
```
Stepping through this function, the error arises in the `upset.CTF()` method, specifically https://github.com/MahShaaban/cRegulome/blob/d5aae84a7a6fc9d8daac3ee2653fa932e30804fb/R/methods.R#L411-L413

Are you able to reproduce this error?

Cheers,
Pete

```R
> devtools::check()
Updating cRegulome documentation
Loading cRegulome
Setting env vars --------------------------------------------------------------------------------------
CFLAGS  : -Wall -pedantic
CXXFLAGS: -Wall -pedantic
Building cRegulome ------------------------------------------------------------------------------------
'/Library/Frameworks/R.framework/Resources/bin/R' --no-site-file --no-environ --no-save --no-restore  \
  --quiet CMD build '/Users/Peter/GitHub/cRegulome' --no-resave-data --no-manual

* checking for file ‘/Users/Peter/GitHub/cRegulome/DESCRIPTION’ ... OK
* preparing ‘cRegulome’:
* creating vignettes ... ERROR
Loading required package: R.oo
Loading required package: R.methodsS3
R.methodsS3 v1.7.1 (2016-02-15) successfully loaded. See ?R.methodsS3 for help.
R.oo v1.21.0 (2016-10-30) successfully loaded. See ?R.oo for help.

Attaching package: 'R.oo'

The following objects are masked from 'package:methods':

    getClasses, getMethods

The following objects are masked from 'package:base':

    attach, detach, gc, load, save

R.utils v2.5.0 (2016-11-07) successfully loaded. See ?R.utils for help.

Attaching package: 'R.utils'

The following object is masked from 'package:utils':

    timestamp

The following objects are masked from 'package:base':

    cat, commandArgs, getOption, inherits, isOpen, parse, warnings

Attaching package: 'dplyr'

The following objects are masked from 'package:stats':

    filter, lag

The following objects are masked from 'package:base':

    intersect, setdiff, setequal, union

Attaching package: 'dbplyr'

The following objects are masked from 'package:dplyr':

    ident, sql

Attaching package: 'tidyr'

The following object is masked from 'package:R.utils':

    extract

Attaching package: 'cRegulome'

The following objects are masked from 'package:graphics':

    hist, plot

The following object is masked from 'package:base':

    print

trying URL 'https://www.dropbox.com/s/t8ga5j8o81jkcuv/test.db.gz?raw=1'
Content type 'application/octet-stream' length 875072 bytes (854 KB)
==================================================
downloaded 854 KB

Joining, by = c("mirna_base", "feature")
Joining, by = c("mirna_base", "feature")
Picking joint bandwidth of 0.0665
Quitting from lines 260-262 (using_cRegulome.Rmd)
Error: processing vignette 'using_cRegulome.Rmd' failed with diagnostics:
Evaluation error: object 'x' not found.
Execution halted
Error: Command failed (1)
```
