### 2017-09-20 #149 @MahShaaban
Hi @PeteHaitch, thanks a lot for taking the time to check the package.
I am unable to reproduce the error.
I can think of one thing, `devtools::check()` overwrites the `NAMESPACE` file which has three manual entries `import(RSQLite)`, `import(R.utils)` and `import(dbplyr)`. I still don't get an error when removing these entries but you may try re-clone the package and run `devtools::check(document=F)` to see if the error goes away.
Have you tried knitting the vignette directly to see if you get the same error?

Thanks again

Mahmoud
