### 2017-09-21 #149 @PeteHaitch
Starting fresh and running `devtools::check(document = FALSE)` still resulted in the same error. The error occurs when I run the vignette code manually, too.

I'll take a closer look tomorrow. In the meantime, can you please post the output of `devtools::session_info()` on the machine you are able to build the vignette so I can compare R version, package versions etc. with what I'm running (I tend to run bleeding edge, which may be causing issues).
