### 2017-09-21 #149 @MahShaaban
I ran `update.packages()` followed by `devtools::check(document=F)`, the test passes with no errors.
Then, I ran the vignette code in the same session followed by `devtools::session_info()`
<summary>Session info</summary>
<pre>
Session info -------------------------------------------------------------------------------
 setting  value
 version  R version 3.4.1 (2017-06-30)
 system   x86_64, darwin15.6.0
 ui       RStudio (1.0.143)
 language (EN)
 collate  en_US.UTF-8
 tz       Asia/Seoul
 date     2017-09-21

Packages -----------------------------------------------------------------------------------
 package        * version date       source
 assertthat       0.2.0   2017-04-11 CRAN (R 3.4.0)
 base           * 3.4.1   2017-07-07 local
 bindr            0.1     2016-11-13 CRAN (R 3.4.0)
 bindrcpp       * 0.2     2017-06-17 CRAN (R 3.4.0)
 bit              1.1-12  2014-04-09 CRAN (R 3.4.0)
 bit64            0.9-7   2017-05-08 CRAN (R 3.4.0)
 blob             1.1.0   2017-06-17 CRAN (R 3.4.0)
 colorspace       1.3-2   2016-12-14 CRAN (R 3.4.0)
 compiler         3.4.1   2017-07-07 local
 cRegulome      * 0.99.0  2017-09-15 Bioconductor
 curl             2.8.1   2017-07-21 CRAN (R 3.4.1)
 datasets       * 3.4.1   2017-07-07 local
 DBI            * 0.7     2017-06-18 CRAN (R 3.4.0)
 dbplyr         * 1.1.0   2017-06-27 CRAN (R 3.4.1)
 devtools         1.13.3  2017-08-02 CRAN (R 3.4.1)
 digest           0.6.12  2017-01-27 CRAN (R 3.4.0)
 dplyr          * 0.7.3   2017-09-09 CRAN (R 3.4.1)
 futile.logger    1.4.3   2016-07-10 CRAN (R 3.4.0)
 futile.options   1.0.0   2010-04-06 CRAN (R 3.4.0)
 ggjoy            0.4.0   2017-09-15 CRAN (R 3.4.1)
 ggplot2        * 2.2.1   2016-12-30 CRAN (R 3.4.0)
 ggridges         0.4.1   2017-09-15 CRAN (R 3.4.1)
 glue             1.1.1   2017-06-21 CRAN (R 3.4.0)
 graphics       * 3.4.1   2017-07-07 local
 grDevices      * 3.4.1   2017-07-07 local
 grid             3.4.1   2017-07-07 local
 gridExtra        2.3     2017-09-09 CRAN (R 3.4.1)
 gtable           0.2.0   2016-02-26 CRAN (R 3.4.0)
 httr             1.3.1   2017-08-20 CRAN (R 3.4.1)
 knitr            1.17    2017-08-10 CRAN (R 3.4.1)
 labeling         0.3     2014-08-23 CRAN (R 3.4.0)
 lambda.r         1.2     2017-09-16 CRAN (R 3.4.1)
 lazyeval         0.2.0   2016-06-12 CRAN (R 3.4.0)
 magrittr         1.5     2014-11-22 CRAN (R 3.4.0)
 memoise          1.1.0   2017-04-21 CRAN (R 3.4.0)
 methods        * 3.4.1   2017-07-07 local
 munsell          0.4.3   2016-02-13 CRAN (R 3.4.0)
 pkgconfig        2.0.1   2017-03-21 CRAN (R 3.4.0)
 plyr             1.8.4   2016-06-08 CRAN (R 3.4.0)
 purrr            0.2.3   2017-08-02 CRAN (R 3.4.1)
 R.methodsS3    * 1.7.1   2016-02-16 CRAN (R 3.4.0)
 R.oo           * 1.21.0  2016-11-01 CRAN (R 3.4.0)
 R.utils        * 2.5.0   2016-11-07 CRAN (R 3.4.0)
 R6               2.2.2   2017-06-17 CRAN (R 3.4.0)
 Rcpp             0.12.12 2017-07-15 CRAN (R 3.4.1)
 reshape2         1.4.2   2016-10-22 CRAN (R 3.4.0)
 rlang            0.1.2   2017-08-09 CRAN (R 3.4.1)
 RSQLite        * 2.0     2017-06-19 CRAN (R 3.4.0)
 rstudioapi       0.7     2017-09-07 CRAN (R 3.4.1)
 scales           0.5.0   2017-08-24 CRAN (R 3.4.1)
 stats          * 3.4.1   2017-07-07 local
 stringi          1.1.5   2017-04-07 CRAN (R 3.4.0)
 stringr          1.2.0   2017-02-18 CRAN (R 3.4.0)
 tibble           1.3.4   2017-08-22 CRAN (R 3.4.1)
 tidyr          * 0.7.1   2017-09-01 CRAN (R 3.4.1)
 tidyselect       0.2.0   2017-08-30 CRAN (R 3.4.1)
 tools            3.4.1   2017-07-07 local
 UpSetR           1.3.3   2017-03-21 CRAN (R 3.4.0)
 utils          * 3.4.1   2017-07-07 local
 VennDiagram      1.6.17  2016-04-18 cran (@1.6.17)
 withr            2.0.0   2017-07-28 CRAN (R 3.4.1)
</pre>
