### 2017-09-21 #149 @PeteHaitch
After downgrading some github-version packages to CRAN/BioC release-version the error has gone away (I suspect it was changes in **rlang** not being compatible with release-versions of other tidyverse pkgs. In any case, this was my problem not that of **cRegulome**). I can now proceed with my review. Thanks for your patience!
