### 2017-10-12 #149 @PeteHaitch
@MahShaaban, @karthik here is my review

>##### For packages co-submitting to JOSS

__Package not being co-submitted to JOSS__

>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

**cRegulome** provides access to a database of pre-computed correlations of transcription factor-gene pairs and microRNA-gene pairs in cancer. It also provides several options for plotting these data.

Unfortunately, what is lacking right now is a clear statement of need and example(s) illustrating the potential use cases of **cRegulome**. The documentation is insufficient for a new user to know when they might make use of **cRegulome** and how it would be useful. The `README` appears to be incomplete (the 'More' and 'Citation' sections, in particular).

How are you planning to distribute the package, @MahShaaban? You didn't tick the box saying you plan to upload to CRAN; is that correct? Are you instead planning a Bioconductor submission? If submitting to Bioconductor then I anticipate that you will require much more integration with existing infrastructure and packages. I'm happy to discuss this further, as I have quite a bit of experience with BioC. If not CRAN or BioC, then how do you plan to distribute the package?

The package claims to support all cancer types studied by the TCGA (a consortium whose data are widely used in bioinformatics/computational biology). However, in testing the package I was only able to successfully query 2/34 cancer types (the 2 types used in the examples were the only ones I could get to work). Here's a reproducible example (I'm not 100% sure this isn't user error on my part):

``` r
library(cRegulome)
#>
#> Attaching package: 'cRegulome'
#> The following objects are masked from 'package:graphics':
#>
#>     hist, plot
#> The following object is masked from 'package:base':
#>
#>     print

get_db(test = FALSE)
R.utils::gunzip('cRegulome.db.gz', overwrite = TRUE)
conn <- DBI::dbConnect(RSQLite::SQLite(), "cRegulome.db")

# All study IDs listed at
# https://tcga-data.nci.nih.gov/docs/publications/tcga/
dat <- get_mir(conn,
               mir = c('hsa-let-7b', 'hsa-mir-134'),
               study = c("LAML", "ACC", "BLCA", "LGG", "BRCA", "CESC", "CHOL",
                         "COAD", "ESCA", "FPPP", "GBM", "HNSC", "KICH", "KIRC",
                         "KIRP", "LIHC", "LUAD", "LUSC", "DLBC", "MESO", "OV",
                         "PAAD", "PCPG", "PRAD", "READ", "SARC", "SKCM", "STAD",
                         "TGCT", "THYM", "THCA", "UCS", "UCEC", "UVM"),
               min_cor = .5,
               max_num = 100,
               targets_only = TRUE)
#> Error: Strings must match column names. Unknown columns: LAML, LGG, BRCA, CESC, CHOL, COAD, ESCA, FPPP, GBM, HNSC, KICH, KIRC, KIRP, LIHC, LUAD, LUSC, DLBC, MESO, OV, PAAD, PCPG, PRAD, READ, SARC, SKCM, STAD, TGCT, THYM, THCA, UCS, UCEC, UVM
```

The SQL databases downloaded by `cRegulome:: get_db()` are hosted on a Dropbox account. This feels somewhat fragile. Also, the URL of the resource is not displayed to the user unless they examine the source code for `cRegulome:: get_db()`. I'm unsure of rOpenSci best practices in this area, but I feel like this could be improved; perhaps @karthik can advise.

Some further questions on the contents of the database:

- Did you create the database @MahShaaban?
- What are the details of how these pre-computed correlations computed? Are the documented somewhere?
- Is `cRegulome.db` simply re-packaging and re-distributing data from [Cistrome Cancer](http://cistrome.org/CistromeCancer/) and [mirCancerdb](https://mahshaaban.shinyapps.io/miRCancerdb/) data? Is that permitted under their respective licenses?

The `get_db()` function could avoid re-downloading the database for each project each time by using a cache. There are several R packages available for doing this, e.g. [BiocFileCache](https://bioconductor.org/packages/BiocFileCache/).

Can users add additional databases? If so, then some guidance should be provided (e.g., format of database).

You should import the pipe operator into the package namespace and not define it inline (e.g. as in https://github.com/MahShaaban/cRegulome/blob/d5aae84a7a6fc9d8daac3ee2653fa932e30804fb/R/methods.R#L119)

Additional questions and comments on specific issues from the rOpenSci review template are below. Hopefully these comments are helpful. Let me know if you have any questions.

#### A statement of need

- It's not clear from the `README` what problems **cRegulome** is designed to solve. I think an example of where **cRegulome** has been useful in your research (in simplified form) might help. This would clarify the target audience. A hypothetical example, "Alice is studying pancreatic cancer and she has a list of genes that seem to be important in this cancer. Now she wants to investigate which miRNA and TFs might be regulating these genes - **cRegulome** can help her do this by X, Y and Z!"
- The README mentions that there are 'many bioinformatics tools ... to predict and identify transcription factors and microRNA targets and their role in development of diseases including cancers'; what is that these tools lack and which **cRegulome** adds or improves upon?
  - In particular, why use **cRegulome** instead of [Cistrome Cancer](http://cistrome.org/CistromeCancer/) or [mirCancerdb](https://mahshaaban.shinyapps.io/miRCancerdb/) (which **cRegulome** builds on)?

#### Installation instructions

- The instructions are clear and work. However, when running `R CMD check` via `devtools::check()` I get a note `Package unavailable to check Rd xrefs: ‘targetscan.Hs.eg.db’`. [**targetscan.Hs.eg.db** is available from Bioconductor](https://bioconductor.org/packages/targetscan.Hs.eg.db/). I'm unsure of rOpenSci's policy regarding packages linked to in docs but not installed as part of the package; @karthik can you please advise.

#### Vignette(s)

- The vignette runs successfully but is focuses too much on the technical/R side and not enough on the practical side. Currently, it mostly repeats what's already in the man pages, documenting the various classes & methods and giving some examples but without any real context or motivation.
- You can go into greater detail here than in the `README`,  writing less about the technical details and more about the scientific context in the than in the man pages. Some of the sort of questions I'd try to address in the vignette are:
  - What sort of experiment/analysis might you be running that would benefit from analysis with cRegulome?
  - What would you do hope to learn from looking at correlations of TFs and microRNAs with your gene list?
  - How do the plots help determine if there's something interesting for a given TF, miRNA or gene?
  - What does a 'good' plot look like? What does a 'bad' plot look like? Basically, how do I know if cRegulome is giving me useful information?

#### Function Documentation

- The man pages are function/method-focused. It may be easier to follow and understand if they were re-written to be class-focused, with most (all?) methods for that class described there together. For example, a man page for the _cTF_ class might include a description of the _cTF_ class, the `cTF()` constructor function, the `print()` method for the _cTF_ class, and perhaps the `tidy()` and various plotting-methods. Similarly, a separate man page for the _cmicroRNA_ class and its associated methods. When defining a new generic, such as `venn.diagram()` and unlike `plot()` or `hist()`, then it is worth writing a man page describing the generic, i.e. laying out the general contract that any methods should adhere to. You might also document the specific `venn.diagram.cTF()` and `venn.diagram.cmicroRNA()` on the same page to illustrate a concrete implementation of the generic.
- 'Description' of `get_mir()` and `get_tf()` are identical - probably worth mentioning in each case what is being accessed from the database
- Not enough detail for `hist`, `joy()`, `plot()`. What is the histogram/joy plot of? Correlations? Something else?  I think more descriptive names than `hist()`, `joy()`, `plot()`, `upset()`, would also help, e.g., `cor_hist()`. Also, `plot()` and `hist()` are S3 generics from the graphics package, but in cRegulome these S3 methods have different signatures to the generic. That's generally a bad idea and a warning sign that you should probably use a different name for this functions, one that is more specific and more descriptive.
- `min_cor` arg of `get_mir()` and `get_tf()` might better be named `min_abs_cor` since it's the "an absoute correlation minimmum"
- `tf` in `get_tf()` is documented as "the official gene symbols of the genes contains the transcription factor". Unfortunately, there are many 'official' gene names/symbols/IDs, so this needs to be clarified. Based on the example for `get_tf()`, which uses `tf = c("AFF4", "ESR1"), these appear to be 'HUGO Gene Symbols' (specified by the HUGO Gene Nomenclature Committee).
- Please run `devtools::spell_check()` as there are many spelling mistakes in the man pages and elsewhere in the documentation

#### Community guidelines

- Contribution guidelines need to be added to the `README` or a `CONTRIBUTING` file added

- As noted above, the package doesn't currently work for 32/34 cancer types with a TCGA study

#### Automated tests

- All tests pass on local machine
- I tested `get_tf()` and `get_mir()` with some nonsense input to check how robust these were to errors
  - `max_num` of is documented to be an integer. One of tests is that `max_num = -1` leads to error `max_num should be an integer between 1 and Inf`. However, `max_num = 0.3` and `max_num = 0` do not result in an error
