### 2017-10-13 #149 @MahShaaban
@PeteHaitch, I'd like to thank you first for your time and effort.
Here is a point-to-point reply to your comments

- I added a paragraph to the `README` and rephrased some of the functions documentation to clarify the use of the package and the functions. In additions, I added a second vignette with a case study using **cRegulome**. Due to the long run time of the vignette, I temporarily have it in the `.Rbuildignore` after checking that it runs locally fine. I am not sure if there is a workaround that.
-  As for the instructions for installing the development version of the package, I assume this is the development version since there are no stable releases yet. Please, let me know if you recommend another way.
- I added a field `Author@R` to the `DESCRIPTION` file and removed `Author` and `Maintainer`

- I am planning to submit to Bioconductor. I looked into using a variation of the `eSet` as a container but at last, I thought the `sqlite` is more convenient. I'd appreciate your guidance on the integration part.
- I am using Dropbox for now, and probably can use figshare or do you suggest other services?
- I built the database and here is a repo explaining the steps and link to the resources used [cRegulomedb](https://github.com/MahShaaban/cRegulomedb). It would be great to get you input on it as well.
- Cistrome is licensed under CC and miRCancer is GPL-2 (by our lab)
- I will look further into that, for now `get_db()` olny download the database if it doesn't exist into the working directory.
- Currently, users can't import additional databases
- The pipe operator is now imported through the namespace
- The package is using **ggridges** now instead of **ggjoy**

**A statement of need**
I think the new paragraph in `README` and the second vignette should address your concerns regarding this point

**Installation instructions**
I am also not sure about packages mentioned in docs!

**Vignette(s)**
Please check the second vignette and let me know if it needs further rewriting.

**Function Documentation**
- I was tempted to do exactly that, document the classes `cTF` and `cmicroRNA` but I read others advise that for **S3** there is no need to document the class and instead the document the constructor functions! Do you think I should use something like `@method` in the man page of the constructor functions to show the methods used with these classes?
- I changed the description of `get_tf()` and `get_mir()` as you suggested.
- I changed the methods names to `cor_hist()`, `cor_plot`, `cor_joy` and `cor_upset` to avoid these conflicts.
- `min_cor` is now `min_abs_cor`
- I specified the HUGO Nomenclature as the gene symbols
- I ran `devtools::spell_check()` and corrected the errors

**Community guidelines**
I added a `CONTRIBUTING` file.

**Functionality**
There was a problem with the urls, so in either case of `get_db(test = TRUE)` and `get_db(test = FALSE)` the function downloaded the same test subset of the database which contains only two studies. I corrected the urls and it should download the full database with `test = FALSE`.  You can check which studies are available using `dbListFields()` in either tables `cor_tf` and `cor_mir`

**Automated tests**
I edited  the code in `get_tf()` and `get_mir()` to handle `max_num = 0.3` and `max_num = 0` as will as test cases for that.

Please, let me know if I missed anything or if you have further concerns
