### 2017-10-16 #149 @PeteHaitch
Hi @MahShaaban

Thank you for your response to the review and the changes to the package. There are still a few to sort through, and I'll address those point-by-point, below.

But first, my main concern, and where I need guidance from @karthik or someone else from rOpenSci, is that rOpenSci and Bioconductor (where you plan to submit your package) have different review criteria, expectations, and requirements (https://www.bioconductor.org/developers/package-submission/, https://www.bioconductor.org/developers/package-guidelines/). I don't speak on behalf of Bioconductor, however, I know that they aim to have packages that interoperate well with existing Bioconductor packages and data structures. Currently, I don't think **cRegulome** does this, which may be a difficulty when submitting to Bioconductor for review. From rOpenSci's perspective this (quite reasonably) may not be a problem, but your package will be reviewed upon submission to Bioconductor and, ultimately, it will have to satisfy Bioconductor's review criteria and not rOpenSci's in order to be accepted as a Bioconductor package.

So right now, I see 3 options:

1. Figuring out a way to satisfy both rOpenSci and Bioconductor review processes (potentially rather tricky and requiring a lot of back-and-forth communication)
2. Aiming to address rOpenSci's review criteria and submitting to CRAN instead of Bioconductor
3. Submitting to Bioconductor and not submitting to rOpenSci

To emphasise, I think **cRegulome** may be a useful bioinformatics package, but I'm unsure how to reconcile the expectations and requirements of a rOpenSci and Bioconductor packages, especially when the package is yet to be reviewed by Bioconductor.

### Point-by-point comments

#### Update to `README` and addition of 'case study' vignette

The updates to the README are good and address the 'statement of need'. The 'case study' vignette is a good start on exactly the type of thing I was looking for as a potential new user for the package. I have a few suggestions for improvements to this vignette.

> Due to the long run time of the vignette, I temporarily have it in the .Rbuildignore after checking that it runs locally fine.

- I think this vignette should be built as part of the package. The long running time of the vignette is a bit of a problem (although I didn't find it unreasonable). However, my bigger concern is that it requires the full `cRegulome.db` to be downloaded (750 Mb). I wonder whether this could be mocked, e.g., by having an unevaluated code chunk showing how to download and set up the database and then a hidden code chunk that loads the relevant subset of the database for use in the vignette. From then on, the the code chunks in the vignette would be evaluated, behaving as if the entire database was available. This would allow the vignette to be built more quickly and requiring fewer resources while still walking the user through the steps required to do replicate the analysis with the full database.
- I removed the vignette from .Rbuildignore and ran `R CMD build` + `R CMD check`. Doing so revealed a few issues with the vignette - these `NOTES` from `R CMD check` will need to be addressed (see below).+

```R
Peter@peters-mbp-2 GitHub $ R CMD check cRegulome_0.0.0.9000.tar.gz

I plan to fix this - the report is just in case I forget or get hit by a bus.
   -- Ross Ihaka (reporting a bug in persp)
      R-help (October 2003)

* using log directory ‘/Users/Peter/GitHub/cRegulome.Rcheck’
* using R version 3.4.1 (2017-06-30)
* using platform: x86_64-apple-darwin15.6.0 (64-bit)
* using session charset: UTF-8
* checking for file ‘cRegulome/DESCRIPTION’ ... OK
* checking extension type ... Package
* this is package ‘cRegulome’ version ‘0.0.0.9000’
* package encoding: UTF-8
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... OK
* checking for portable file names ... OK
* checking for sufficient/correct file permissions ... OK
* checking whether package ‘cRegulome’ can be installed ... OK
* checking installed package size ... NOTE
  installed size is  6.9Mb
  sub-directories of 1Mb or more:
    doc       5.1Mb
    extdata   1.7Mb
* checking package directory ... OK
* checking ‘build’ directory ... OK
* checking top-level files ... OK
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... OK
* checking R files for non-ASCII characters ... OK
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... OK
* checking whether the package can be loaded with stated dependencies ... OK
* checking whether the package can be unloaded cleanly ... OK
* checking whether the namespace can be loaded with stated dependencies ... OK
* checking whether the namespace can be unloaded cleanly ... OK
* checking dependencies in R code ... NOTE
Namespaces in Imports field not imported from:
  ‘R.utils’ ‘RSQLite’ ‘dbplyr’
  All declared Imports should be used.
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... OK
* checking Rd files ... OK
* checking Rd metadata ... OK
* checking Rd cross-references ... OK
* checking for missing documentation entries ... OK
* checking for code/documentation mismatches ... OK
* checking Rd \usage sections ... OK
* checking Rd contents ... OK
* checking for unstated dependencies in examples ... OK
* checking installed files from ‘inst/doc’ ... OK
* checking files in ‘vignettes’ ... OK
* checking examples ... OK
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ...
  Running ‘testthat.R’
 OK
* checking for unstated dependencies in vignettes ... NOTE
'::' or ':::' import not declared from: ‘AnnotationDbi’
'library' or 'require' calls not declared from:
  ‘clusterProfiler’ ‘igraph’ ‘org.Hs.eg.db’ ‘readxl’ ‘stringr’
  ‘tidyverse’
* checking package vignettes in ‘inst/doc’ ... OK
* checking running R code from vignettes ...
   ‘case_study.Rmd’ using ‘UTF-8’ ... OK
   ‘using_cRegulome.Rmd’ using ‘UTF-8’ ... OK
 NONE
* checking re-building of vignette outputs ... OK
* checking PDF version of manual ... OK
* DONE

Status: 3 NOTEs
See
  ‘/Users/Peter/GitHub/cRegulome.Rcheck/00check.log’
for details.
```

- The 'case study' vignette needs to be spell-checked and proof-read
- It looks like you are trying to use **BiocStyle** to format the 'case study' vignette. However, this is not being picked up because you have another YAML chunk ahead of the one that sets up **BiocStyle**. There should only be one YAML chunk and it needs to be the one referencing **BiocStyle**, I believe.
- Why are different 'study' terms used in the `get_tf()` and `get_mir`() examples in the 'case study' vignette (see below)?

```R
creg_tf <- get_tf(conn,
                  tf = unique(tf$SOURCE),
                  study = 'STES*')

creg_mir <- get_mir(conn,
                    mir = tolower(unique(mir$AccID)),
                    study = 'STES')
```

- The 'case study' vignette would benefit from including more explanation/interpretation of the results (e.g., when you plot correlations for specific TFs and microRNAs from the STES data). This is especially true for the more complicated figures, like the network diagram created with **iGraph** and the clustering dendogram.
- It would be good to have a function that converts the **cRegulome** output to the **igraph** input. Although documented in the vignette, the user currently has to do a bit of data manipulation, a process which could be turned into a function and included as part of **cRegulome**.

#### Bioconductor integration

> I am planning to submit to Bioconductor. I looked into using a variation of the eSet as a container but at last, I thought the sqlite is more convenient. I'd appreciate your guidance on the integration part.

- This obviously depends on whether you submit to Bioconductor and would need to be addressed to the reviewer of your Bioconductor submission. For what it's worth, a tidy data frame (tibble or otherwise) seems a reasonable data structure, to me.
- Traditionally, Bioconductor has made heavy use of the [AnnotationDBI](https://www.bioconductor.org/packages/AnnotationDbi/) package to interface with annotation data packages using SQLite data storage.
- A more modern, 'tidyverse'-styled approach has recently been developed with the [Organism.dplyr](http://bioconductor.org/packages/Organism.dplyr/) package. This style might better suit **cRegulome**.

#### Database

> I am using Dropbox for now, and probably can use figshare or do you suggest other services?

- figshare seems a good option, @karthik may have other suggestions. If submitting to Bioconductor there is specific infrastructure set up for this, such as http://bioconductor.org/packages/AnnotationHub/ and http://bioconductor.org/packages/ExperimentHub/
- The size of the `cRegulome.db` is on the larger size (750 Mb).
	- As suggested in my initial review, it would good to use a caching solution so that this doesn't need to be re-downloaded for each project
	- Would it be possible to host the database somewhere and allow users to query it, only downloading the data they need for their analysis?

> I built the database and here is a repo explaining the steps and link to the resources used cRegulomedb.

- This should be linked to in the **cRegulome** documentation.
- The links in the 'More' section of https://github.com/MahShaaban/cRegulomedb/blob/master/README.md don't work

#### Licensing

- Could you please link to the license for Cistrome Cancer and miRCancer

#### Function Documentation

> I was tempted to do exactly that, document the classes cTF and cmicroRNA but I read others advise that for S3 there is no need to document the class and instead the document the constructor functions! Do you think I should use something like @method in the man page of the constructor functions to show the methods used with these classes?

Again, this comes down to style differences, here between Bioconductor/S4 and tidyverse/S3. This choice will depend on whether you submit to Bioconductor or CRAN, I think

#### Community guidelines

Please include links to the **issues** and **pull requests** pages. You might also like to use a `CONTRIBUTING.md` file from other rOpenSci projects as a template (e.g., https://github.com/ropensci/spocc/blob/master/.github/CONTRIBUTING.md, https://github.com/ropensci/taxize/blob/master/.github/CONTRIBUTING.md, https://github.com/ropensci/bold/blob/master/.github/CONTRIBUTING.md)

#### Functionality

- With the updated database URL there are still 4 cancer types without any data. Why is that? Should they be available in the database? At least one of these, 'GBM'/'GBMLGG', seems to be due to a potential switch in IDs used for cancer types.

``` r
library(cRegulome)
#>
#> Attaching package: 'cRegulome'
#> The following object is masked from 'package:base':
#>
#>     print
get_db(test = FALSE)
R.utils::gunzip('cRegulome.db.gz', overwrite = TRUE)
conn <- DBI::dbConnect(RSQLite::SQLite(), "cRegulome.db")

# All study IDs listed at
# https://tcga-data.nci.nih.gov/docs/publications/tcga/
dat <- get_mir(conn,
               mir = c('hsa-let-7b', 'hsa-mir-134'),
               study = c("LAML", "ACC", "BLCA", "LGG", "BRCA", "CESC", "CHOL",
                         "COAD", "ESCA", "FPPP", "GBM", "HNSC", "KICH", "KIRC",
                         "KIRP", "LIHC", "LUAD", "LUSC", "DLBC", "MESO", "OV",
                         "PAAD", "PCPG", "PRAD", "READ", "SARC", "SKCM", "STAD",
                         "TGCT", "THYM", "THCA", "UCS", "UCEC", "UVM"),
               min_abs_cor = .5,
               max_num = 100,
               targets_only = TRUE)
#> Error: Strings must match column names. Unknown columns: LAML, FPPP, GBM, MESO

# Manually list cancer types in database
DBI::dbListFields(conn, "cor_mir")
#>  [1] "feature"    "mirna_base" "ACC"        "BLCA"       "BRCA"       "CESC"       "CHOL"
#>  [8] "COAD"       "COADREAD"   "DLBC"       "ESCA"       "GBMLGG"     "HNSC"       "KICH"
#> [15] "KIPAN"      "KIRC"       "KIRP"       "LGG"        "LIHC"       "LUAD"       "LUSC"
#> [22] "OV"         "PAAD"       "PCPG"       "PRAD"       "READ"       "SARC"       "SKCM"
#> [29] "STAD"       "STES"       "TGCT"       "THCA"       "THYM"       "UCEC"       "UCS"
#> [36] "UVM"
```

#### Other

- The default print method is masked when loading **cRegulome**. This is generally bad idea. Your S3 `print()` methods, `print.cmicroRNA()` and `print.cTF()` should have the same signature as `base::print()`, i.e. `print(x, ...)` rather than `print(ob, ...)`

``` r
library(cRegulome)
#>
#> Attaching package: 'cRegulome'
#> The following object is masked from 'package:base':
#>
#>     print
```
