### 2017-10-16 #149 @mmulvahill
- [x ] **Vignette(s)** demonstrating major functionality that runs successfully locally

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Forgive any inaccurate use of terminology.  I've developed databse API's for microRNA data, but my work as a statistician and analyst is in a different area of biomedical research.

#### **A statement of need**

Package is dual purpose -- 1. serving data for microRNA annotations and transcription factor correlations and 2. provide a set of descriptive methods and visualizations for these data types.  This is clear enough from the README and vignette.

My preference is to keep the readme simple -- installation instructions and brief statement of purpose -- and provide detailed instructions and scientific context in the vignettes.  Along these lines, I would shorten the readme and add to the vignette.

Bioconductor has specific classes and package templates for packages that serve to annotation data.  (I just went through this for our `multiMiR` package, which provides functions for querying a collection of 14 miRNA databases.  While I wasn't required to implement the full `AnnotationDbi` structure, I did have to implement the data access and filter functions -- `AnnotationDbi::select()`, `AnnotationDbi::keys()`, `AnnotationDbi::columns()`, and `AnnotationDbi::keytypes()`).

Also, I would consider separating it into two packages, especially if the miRNA/tf correlation statistics will have other applications to other databases and if the cRegulome.db (or Cistrome and miRCancer) database is not yet on Bioconductor (I didn't see the component DBs in an initial, cursory look).

#### **Installation instructions:** for the development version of package and any non-standard dependencies in README

Installed fine for me.  A few notes from `devtools` --

From `devtools::document()`:

- Looks like the package `dbplyr` is only used in the vignette and manuscript
  -- consider moving to Suggests from Imports.
- The `R.utils::gunzip` function is used in many places (vignette, manuscript,
  unrun example), but is not actually used in the package code. Consider moving
  to suggests or adding a roxygen2 line to import the function (in the
  `get_tf()` documentation: `#' @importFrom R.utils gunzip`.
- `RSQLite::SQLite()` has the same situation as `gunzip` above, except that it
  is used in examples executed with `devtools::check()`.  Consider adding an
  `#' @importFrom RSQLite SQLite` line to documentation of functions that use
  this function in examples.

Either of the `@importFrom` suggestions could be `@import <pkgname>` also,
though the former is preferable to avoid function name conflicts and reduce namespace clutter.

From `devtools::check()`:
- Add .git/ to the .Rbuildignore
- "Package unavailable to check Rd xrefs: 'targetscan.Hs.eg.db' -- resolves if
  targetscan.hs.eg.db is installed from bioconductor (doesn't have to be
  loaded). Doesn't affect installation.  Not sure the best solution to this one.

From `devtools::spell_check()`:
- Looks like some spelling errors still exist in the readme and at least one instance of 'to' instead of 'two' in the 'Database query' section of the vignette, but otherwise most pure spelling errors were resolved with recent commits.

#### **Vignette(s)** demonstrating major functionality that runs successfully locally

Runs successfully and demonstrates functionality.  Suggestions:

- Consider wrapping the database connection functionality into the package.  `get_db()` could just as easily test for the existence of `cRegulome.db` and return a `dbConnect` connection to it. Requiring the user to understand DBI, RSQLite, R.utils::gunzip on top of figuring out your package seems like a tall order.  Before making any changes though, look into Bioconductor's DB utilities and S4 classes.
- At >2GB, cRegulome.db is too large to include in the package itself.  Look into database hosting options.  We host our `multiMiR` database on our university servers, but there are likely other cheap or free options.
- `venn.diagram` overlays on top of `cor_joy` plot (line 250).  If unintentional, add `dev.off()` somewhere to reset the plotting device.

#### **Function Documentation:** for all exported functions in R help

- Function documentation overall looks good.
- The naming scheme `as.<s3 class name>` is typical of s3 constructors -- consider using this, but again Bioconductor integration may require changing to S4 classes.
- More details on the methods (citations and help page descriptions) would be helpful.

#### Function naming

- Consider renaming `cRegulome::tidy` or importing and overloading `broom::tidy`

Overall the biggest thing to consider is Bioconductor integration and whether the database and its summary functions are independently novel enough to be better suited to two packages.  On the latter point, I can see an argument for both approaches.
