### 2017-10-19 #149 @MahShaaban
Hi @PeteHaitch,

I think at this point I would go with the first option, submit to rOpenSci and CRAN. Integrating the package in the Bioconductor seems like it would need substantial changes that wouldn't be natural based on the choices I already made.

* I followed your advice in adding the data needed for running the second vignette as a part of the package so to avoid downloading the whole database.

*  I fixed the notes that appeared in the `R CMD build + check` by adding the vignette packages to the `Suggests` section in the `DESCRIPTION` file.

* The note about the size of the package persists, even after changing the compression of the included data and the vignette graphs. I am not sure what goes in the `doc` folder! I hope you have some advice for me here.

* I edited the vignette and added some explanations to the graphs in the context they are used.

* I moved the database to **figshare** and updated the urls.

* I linked the repo that contains the details and the scripts for building the database.

* I edited the `CONTRIBUTING` file using the templates you suggested.

* I used the TCGA names as provided by the Cistrome Cancer and the `RTCGA` package when building the database. The different naming might be the reason for the missing studies but I also know that not all studies are included in the Cistrome Cancer.

* I changed the name of the`print` method to `cor_print` to avoid the conflict.
