### 2017-10-19 #149 @MahShaaban
Hi @mmulvahill,
Thanks for taking the time to review the package. Here are my reply and the changes I made based on your comments.

* I reordered the packages between `Suggests` and `Imports` as you suggested.
* I added `./.git` to `.Rbuildignore`.
* I added `targetscan.Hs.eg.db` to `Suggests` for now.
* I ran `devtools::spell_check()` and fixed errors.
* I originally wrote `get_mir` and `get_tf` to connect to the database directly but it was harder to test and debug. So I prefer to keep connecting to and decompressing the database file separate and obvious to the user.
* Now the database is on **figshare**
* I added `dev.off()` as you suggested.
* I actually thought it's a typical practice to name the S3 class and its constructor by the same name. Please, correct me if I am wrong and I would change them.
* I changed the function name `tidy` to `cor_tidy`.

Please let me know if you have further comments.
