### 2017-10-22 #149 @PeteHaitch
Hi @MahShaaban, those changes sound really promising. Thanks!

I'm travelling currently and at a conference this coming week - I hope to have time to take a proper look through those changes but it may not happen. I appreciate your patience
