### 2017-11-03 #149 @mmulvahill
@MahShaaban Thanks for making those changes.

Installation isn't quite yet as error-free as it probably should be.   If I use docker (rocker/verse) to start fresh, I can install with `devtools::install_github` without error, but this doesn't build/provide the vignettes.

If I download and install, building vignettes:

```{r}
download.file("https://github.com/MahShaaban/cRegulome/archive/master.zip", "cRegulome.zip)
unzip("cRegulome.zip")
file.rename("cRegulome-master/", "cRegulome")
devtools::install('cRegulome', build_vignettes = TRUE)
```
It fails due to missing packages.  If you are not submitting to Bioconductor, it's probably a good idea to use rmarkdown::html_vignette instead of BiocStyle.  For the other Bioconductor packages, including code for installing them within the vignette may make this process smoother.  i.e.

```{r}
if (!require(BiocInstaller)) source("https://bioconductor.org/biocLite.R")
if (!require(AnnotationDbi)) biocLite("AnnotationDbi")
if (!require(org.Hs.eg.d))   biocLite("org.Hs.eg.db")
if (!require(clusterProfiler)) install.packages("clusterProfiler")
```

There good guidance out there (http://r-pkgs.had.co.nz/description.html#dependencies), but I haven't exactly figured out how CRAN and install.packages() handles vignettes and packages in `Suggests: `  @PeteHaitch  or @karthik may be able to add some detail here.

On the S3 class constructors, I think your right.  It seems that the `as.<class name>` structure is more for forcing one type to another.
