### 2017-11-06 #149 @MahShaaban
Thanks, @mmulvahill for your review.
Regarding the missing packages for building the vignette, I am not sure I am supposed to include code to install them when missing either these from bioconductor or CRAN! Otherwise why they should go in `Suggests:` in the first place! I don't mind adding this chunk `if (!require(package)) biocLite('package')`, but wouldn't it make sense in this case to just move them to `Imports`?
I'd like to know your thoughts on this.
