### 2017-11-13 #149 @mmulvahill
My preference is that optional packages be placed in `Suggests:` (not installed by default), but that optional code (as in examples, vignettes, and optional features of a fuction) handle the installation of the optional packages when the user requests it.

Minimizing dependencies improves the reliabiltiy of your code.  `Suggests:` combined with `if (!require(package)) install.packages(<pkg>)` lets you show extensions to your package without forcing more packages, dependencies, and complexity on the user while preventing errors from being thrown if and when the user tries out these extensions.

It's not a requirement for CRAN or Bioconductor, but it is allowed and improves the user's experience of your tool.
