### 2017-11-14 #149 @MahShaaban
Thanks, @mmulvahill for the explanation I currently have 12 packages in the **Suggests** field of the `DESCRIPTION` file. Some of these packages are used in examples and vignettes.

```
Suggests: knitr,
    rmarkdown,
    testthat,
    covr,
    dbplyr,
    R.utils,
    clusterProfiler,
    igraph,
    org.Hs.eg.db,
    readxl,
    stringr,
    AnnotationDbi
```
If I understand you correctly, I need to add that optional code `if(!require(package)) install.packages(<pkg>)` or the equivalent for Bioconductor packages every time one of these packages is used. Right?
