### 2017-11-15 #149 @MahShaaban
Here, I added this chunk to the vignette to download the missing packages form their respective sources, Bioconductor or CRAN in the last commit
Now, I get the following note with running `devtools::check()`
`'library' or 'require' call not declared from:  ‘BiocInstaller’`
