### 2017-11-16 #149 @PeteHaitch
Looking into this, @MahShaaban. One thing, **clusterProfiler** is a Bioconductor package and so should be installed with `BiocInstaller::biocLite()` rather than `install.packages()` in the vignette
