### 2017-11-16 #149 @PeteHaitch
Also, the section in the vignette with the installation of suggested packages should be in a `include=TRUE` chunk otherwise the reader won't see it in the rendered output
