### 2017-11-17 #149 @MahShaaban
Thanks @PeteHaitch for spotting these issues. I removed the `include=FALSE` and changed **clusterProfiler** to be installed using `biocLite()`. However, I still get the same note `library' or 'require' call not declared from: ‘BiocInstaller`!
