### 2017-11-17 #149 @PeteHaitch
Adding **BiocInstaller** to `Suggests` in the `DESCRIPTION` fixes this (for reference, https://stackoverflow.com/questions/15648772/how-do-i-prevent-r-library-or-require-calls-not-declared-warnings-when-dev)
