### 2017-12-20 #149 @mmulvahill
I haven't been able to get the package to install and build vignettes on a clean install (hence my delay responding).  This appears to be due to a quirk in automating the bioconductor installs that I can't quite figure out.

In the 2nd update I link to below, I added some code for loading and updating with biocLite.  Despite this, I still get an error with `devtools::install("./cRegulome", build_vignettes = TRUE)`:

`Error processing vignette 'case_study.Rmd' failed with diagnostics:
object 'org.Hs.eg.db' not found`

Separately, I would add `if (!require()) install` code for R.utils to `using_cRegulome.Rmd`.

[update #1 (to using_cRegulome.Rmd)](https://github.com/mmulvahill/cRegulome/commit/a33dfbb6030be80bdceb4180a8b28e917feda8a8)
[update #2 (to case_study.Rmd)](https://github.com/mmulvahill/cRegulome/commit/2d43cc591a24d86121dd36f85f5e836969a67b66)
