### 2017-12-24 #149 @PeteHaitch
Hi @MahShaaban

The vignette still needs more careful proofreading. In my opinion, the vignettes should be the most useful parts of the package but they can be hard to follow in their current state.

For example in the 'Case Study' vignette, "In the previous section, we showed the code for obtaining the TF/microRNA-gene expression correlation in stomach and esophogeal cancer using Cistrome and miRCancerdb for comparison purpose." But it's not clear to me that this is indeed what was done in the previous section. There are also many typos; try running `spelling::spell_check_package(vignettes = TRUE)`. Please take the time to carefully proofread both vignettes.

It should be made clear in the vignettes that some of what the user is viewing is based on the 'test' database. Otherwise they may be surprised that the output doesn't match their own when they run the code themselves using the full database. Even better may be to use the mocking strategy I suggested in https://github.com/ropensci/onboarding/issues/149#issuecomment-337022051 This shouldn't be too hard to set up.

As I noted in my initial review (https://github.com/ropensci/onboarding/issues/149#issuecomment-336006288):

> The `get_db()` function could avoid re-downloading the database for each project each time by using a cache. There are several R packages available for doing this, e.g. [BiocFileCache](https://bioconductor.org/packages/BiocFileCache/).

I still think this is a really good idea, but I'm willing to move on at this point.

Some minor points:

- Could `get_db()` also take care of unzipping the database? This would simplify the setup process.
- In the 'Case Study' vignette, `conn <- dbConnect(SQLite(), '~/microRNA/cRegulomedb/cRegulome.db')` is not portable since it references a path on your own computer

Cheers,
Pete
