### 2017-12-28 #149 @MahShaaban
Thanks @mmulvahill & @PeteHaitch for these comments.
I will work on them and follow with a commit soon.
Happy holidays
Mahmoud
