### 2018-01-04 #149 @MahShaaban
Here are a few changes I made to address the issues with the installation and the vignette.

## Code changes

- `get_db` now handles the unzipping process. So I moved `R.utils` to the **Imports** section in the `DESCRIPTION` file. This solved the problem of the installation. I successfully installed the package on **rocker/verse** by either `devtools::install_github('MahShaaban/cRegulome', build_vignette=TRUE)` and locally `devtools::install(build_vignette=TRUE)`
- I looked the `BiocFileCache` but couldn't figure out a solution for using it in `get_db`, at least for now

## Vignettes

- I fixed the line `conn <- dbConnect(SQLite(), '~/microRNA/cRegulomedb/cRegulome.db')` to `conn <- dbConnect(SQLite(), './cRegulome.db')`
-  I reviewed it and ran the `spelling::spell_check_package`. Finally, I mentioned the fact that the vignette is using a test set early on

I hope these changes solve the issues raised by @mmulvahill and @PeteHaitch
