### 2018-01-30 #149 @PeteHaitch
I submitted a pull request with fixing the vignette metadata to ensure it's built and installed as part of `R CMD build/check` (also fixes some typos) https://github.com/MahShaaban/cRegulome/pull/1

I tried running through the "Using cRegulome" vignette as a new user might, but it unfortunately does not work. Specifically, if you build the vignette and try running through the steps at https://github.com/MahShaaban/cRegulome/blob/master/vignettes/using_cRegulome.Rmd#getting-started, you'll get an error:

```r
> # load required libraries
> library(cRegulome)
> library(RSQLite)
> library(ggplot2)
> # download the db file when using it for the first time
> if(!file.exists('cRegulome.db')) {
+     get_db(test = TRUE)
+ }
>
> # connect to the db file
> conn <- dbConnect(SQLite(), 'cRegulome.db')
> # enter a custom query with different arguments
> dat <- get_mir(conn,
+                mir = 'hsa-let-7g',
+                study = 'STES',
+                min_abs_cor = .3,
+                max_num = 5)
Error: Unknown column `STES`
```

This is because the test database you connect to as a vignette user (in the `prepare database` chunk with `eval = FALSE`) is not the same database as that used by the vignette itself (in the `connect_db` chunk with `include = FALSE`). I can't link to the exact lines because of how GitHub renders `.Rmd` files, but that should be enough to identify the problem. You need to ensure these are the same database for this type of 'mock' database download/connection to work properly.

You can try building and installing the package with vignettes (`devtools::install_github('MahShaaban/cRegulome', build_vignettes = TRUE)` and then following the build vignette as a new user might.

Incidentally, you may wish to update the installation instructions in the `README.md` to use `devtools::install_github('MahShaaban/cRegulome', build_vignettes = TRUE)` to ensure the user gets the vignettes (these will be build and installed automatically when the package is on CRAN).
