### 2018-01-30 #149 @MahShaaban
Thanks @PeteHaitch for the fixes, very helpful as always.
I added the line `devtools::install_github('MahShaaban/cRegulome', build_vignettes = TRUE)` to the read me as you suggested.
I updated the file and the url so that `get_db(test = TRUE)` would download the exact file that is used in the package and test running the vignettes and both chunks `prepare dataset` and `connect_db` connect to identical files.
