### 2018-02-03 #149 @PeteHaitch
Hi @MahShaaban,

Thanks for the changes. I was able to work through the vignette like a new user.

I have gone through the rOpenSci package review checklist and I am happy to recommend approving **cRegulome** :tada:

Below are a few cosmetic suggestions/changes (some with PRs) that you might consider:

- `cor_print()` could just be print() (https://github.com/MahShaaban/cRegulome/pull/3)
- `venn.diagram()` -> `cor_venn_diagram()` for consistency with other generics (https://github.com/MahShaaban/cRegulome/pull/4)
- `cor_plot()` might scale point size in (-1, 1) rather than relying on the range of the input data. For the example in the vignette, visual difference exaggerates actual difference in correlation values
- Vignette text and code might be out of sync?
  - In the 'Pathway enrichment analysis' section:
	- "Clusters 1 and 3" seems to be clusters 3 and 4?
	- Why no KEGG enrichment results for other clusters?
	- "enrichment of 13 and 7 KEGG pathways respectively" doesn't seem to match the numbers in the figure
	- You might derive the numbers in the text from data to ensure they match the output using [inline R code chunks](http://rmarkdown.rstudio.com/lesson-4.html)
- "We didn’t attempt to replicate the same findings but nevertheless we arrived at similar conclusions."
	- It's not clear to me what the conclusions your example analysis are or how they compare to Shi et al.'s. You might summarise each in the final section of the vignette
