### 2018-02-04 #149 @MahShaaban
Thanks @PeteHaitch for these pull requests and the comments.

- I merged the three requests you made, they certainly improve the consistency
- I made changes in the vignette to sync the text with the graphs
- The function `clusterProfiler::compareCluster()` returns data for the clusters that has significant enrichment hits only. In this case the 3 other clusters don't have any
- In the **motivation** section there was some mistakes probably due to a previous editing. The author's conclusion was that the paper presented a proof of concept for using the TF-microRNA co-regulation networks to identify critical regulators in cancer
- Finally, I experimented with changing the scale of the points in `cor_plot`. I tried using `ggplot2::scale_size(range = c(-1,1))`, the resulting points were very small and not any easier to compare. changing the `range` resulted in almost similar sizes as without using `scale_size` at all!

I really appreciate the time and effort you put in reviewing the package. The comments and the changes you made helped a lot.

Thanks again

Mahmoud
