### 2018-02-05 #149 @mmulvahill
I second @PeteHaitch's recommendation of approving `cRegulome`!  The package, including both vignettes, now builds, installs, and runs successfully for me on a fresh system.  Thanks for your responsiveness @MahShaaban.

- [x ] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [x ] **Vignette(s)** demonstrating major functionality that runs successfully locally

>##### For packages co-submitting to JOSS
>**Not applicable**
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
