### 2018-02-07 #149 @karthik
@MahShaaban Your badge should update soon to peer-reviewed. Congrats on your package being accepted! 🎉

Here are your next steps:

Please also add a footer to the bottom of your README

```
[![](http://www.ropensci.org/public_images/github_footer.png)](http://ropensci.org)
```

- [ ] Please accept my invitation to join a team on the `ropensci` github org. This will allow you to transfer the repo. Once it's transferred, I'll give you write access there so you can update the CI badges.
- [ ]  ~Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)~

Once moved, please re-run all checks in preparation for submission to CRAN. I can help with this if you run into any issues.

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
