### 2018-02-07 #149 @stefaniebutland
Hello @MahShaaban and congratulations on your package being accepted! I'm rOpenSci's Community Manager. As Karthik said, we'd love to have a blog post about `cRegulome`. Here is some technical and editorial info: https://github.com/ropensci/roweb2#contributing-a-blog-post. Let me know if you're interested. Happy to help as needed.
