### 2018-02-11 #149 @maelle
@MahShaaban I see that you've now transferred the repo, fantastic! Welcome! 🎉

* I've activated Appveyor again, the badge is `[![Build status](https://ci.appveyor.com/api/projects/status/gcmojtcsyt7rcwtk?svg=true)](https://ci.appveyor.com/project/ropensci/cregulome)`

* I'm closing this since the package was approved, but feel free to comment here (or ping Stefanie elsewhere) when you're ready to write a blog post. 😺
