### 2018-02-11 #149 @MahShaaban
Thanks @maelle. After I transferred the repo I updated the the ci links including that for Appveyor. It seems to be working; builds for the commits and the badge links to the correct page. Do I need to update the badge again with the link you provided in the previous comment?
